Fix error when reporting error
The initial conditional `(!error && response.statusCode === 200)` can
return false either because of an error or a non-200 status code.

That is fine, but the `else` logic doesn't check `response` was
defined before using. If there was an `error`, `response` will be undefined.

This updates the logic to check for an `error` first, then check for
200 status code, and report it as an error if it isn't.